refactor(backend): extract semantic search to SemanticRepository
